server/requirements.txt
Django>=5.0,<6
djangorestframework>=3.14,<4
django-cors-headers>=4.0,<5
python-dotenv>=1.0
psycopg[binary]>=3.1,<4
djangorestframework-simplejwt>=5.3,<6
Pillow>=10,<11
opensearch-py>=2.4,<3
celery>=5.3,<6
redis>=5,<6
drf-spectacular>=0.27,<1
drf-spectacular-sidecar>=2024.4.1
boto3>=1.34,<2
